Implement delete todo list

+++ lib/src/domain/use_cases/local_use_cases/delete_todo_use_case.dart
@@ -13,6 +13,12 @@ class DeleteTodoUseCase {
 
   Future<Either<Failure, void>> execute(int id) async {
     try {
+      final todo = await localPersistenceRepository.getTodo(id);
+
+      if (todo == null) {
+        return Left(CannotUpdateNullValueFailure("todo is null"));
+      }
+
       await localPersistenceRepository.deleteTodo(id);
 
       return Future.value(Right(unit));
